+++ src/types/Status.ts
@@ -1,0 +1,5 @@
+export enum Status {
+  All = 'all',
+  Active = 'active',
+  Completed = 'completed',
+}
